Demonstrate importing private key and certificate into KeyChain
This code is for proof-of-concept demonstration
purposes only and is not intended for use in a production environment.

estEnroll() generates a keypair using the specified algorithm, keysize, and
alias, and uses the Enrollment over Secure Transport (EST) protocol (IETF RFC 7030)
to enroll for a certificate from Cisco's EST test server (http://testrfc7030.cisco.com).
The private key and enrolled certificate are then placed in the keychain. We use Cisco's sample EST
client found at https://github.com/cisco/libest/tree/master/example/client . We compiled
it (for ARM) and embedded the binary into this application - a cleaner approach for
future work would be for our application to call Cisco's libest through JNI
rather than run an external binary.